+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/tree-base-memo-node/tree-base-memo-node.tsx
@@ -2,8 +2,6 @@ import { memo } from 'react';
 import type { Node, NodeProps } from '@xyflow/react';
 import { Handle, Position } from '@xyflow/react';
 
-import TreeCustomHandle from '../tree-custom-handle/tree-custom-handle';
-
 import * as styles from './tree-base-memo-node.css';
 
 type EmptyNodeDataTypes = Record<string, never>;
@@ -23,30 +21,16 @@ const TreeBaseMemoNode = ({
         type="source"
         position={Position.Right}
         isConnectable={isConnectable}
-        style={{
-          background: 'none',
-          border: 'none',
-          width: 'min-content',
-          height: 'min-content',
-        }}
-      >
-        <TreeCustomHandle labelColor={'grey'} isBaseMemo={true} />
-      </Handle>
+        className={styles.handle}
+      />
 
       <Handle
         id="baseBottom"
         type="source"
         position={Position.Bottom}
         isConnectable={isConnectable}
-        style={{
-          background: 'none',
-          border: 'none',
-          width: 'min-content',
-          height: 'min-content',
-        }}
-      >
-        <TreeCustomHandle labelColor={'grey'} isBaseMemo={true} />
-      </Handle>
+        className={styles.handle}
+      />
     </div>
   );
 };
